Simplify Simulator component by moving some actions to SimulatorUtils
